// src/types/types.ts
export interface AnswerInterface {
  is_correct: any;
  uid: string; //UID
  answer: string; // Answer with a max length of 1000 and min length of 1
  is_answer: boolean; // Is answer flag
}

export interface QuestionInterface {
  uid: string; // ID
  question: string; // Question with a max length of 1000 and min length of 1
  image?: string | null; // Optional Image
  points: number; // Points
  timer: number; // Timer
  answers: AnswerInterface[]; // Array of Answer objects
  is_answered?: boolean; // Is answered flag
  selected_answer?: AnswerInterface; // Optional Selected answer
  duration?: number; // Duration
}

export interface QuestionSetInterface {
  uid: string; // UID
  category_name: string; // Category name with a max length of 300 and min length of 1
  // play_date: string; // Play date-time
  start_datetime: Date; // Start time
  end_datetime: Date; // End time
  status: "PL" | "NP"; // PL - Played, NP - Not Played
  is_payable: "PA" | "NP"; // PA - Payable, NP - Not Payable
  amount_payable: number; // Amount payable
  prize_won: string; // Prize won
  questions: QuestionInterface[]; // Array of Question objects,
  theme_image: string; // Theme image
}

export interface OrganizationInterface {
  uid: string; // UID
  organization_name: string; // Organization name with a max length of 300 and min length of 1
  org_logo?: string | null; // Optional Organization logo
  question_sets: QuestionSetInterface[]; // Array of QuestionSet objects
}

export interface QuestionPlayersInterface {
  id: number; // ID
  question_set: QuestionSetInterface; // QuestionSet object
  timer: number; // Timer
}
// ...
export interface QuizStoreInterface {
  questionPlayers: QuestionPlayersInterface[];
  currentQuestionPlayer: QuestionPlayersInterface | null;
  currentQuestion: QuestionInterface | null;
  currentQuestionIndex: number;
  currentQuestionSet: QuestionSetInterface | null;
  currentOrganization: OrganizationInterface | null;
  organizations: OrganizationInterface[];
  counter: number;
  timer: number;
  isPlaying: boolean;
  questions: QuestionInterface[];
  scoreSummary: QuizScoreSummaryInterface | null;
  currentAnswers: AnswerInterface[];
}

export interface QuizScoreSummaryInterface {
  questionSetUid: string;
  organizationUid: string;
  score: number;
  totalQuestions: number;
  totalCorrectAnswers: number;
  totalWrongAnswers: number;
  totalSkippedAnswers: number;
  totalPoints: number;
  totalDuration: number;
  avgDuration: number;
}